xfer: move Buffer to own file; update comment
overlay: mutex for Weave status
